semantic-layer/pyproject.toml
[project]
name = "semantic-layer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.0.2",
    "ibis-framework[duckdb]>=11.0.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scipy>=1.16.3",
]
